+++ src/Element_TEST.cc
@@ -296,14 +296,33 @@ TEST(Element, ToStringElements)
   sdf::ElementPtr parent = std::make_shared<sdf::Element>();
   sdf::ElementPtr child = std::make_shared<sdf::Element>();
 
+  parent->SetName("parent");
+  child->SetName("child");
+
   parent->InsertElement(child);
   ASSERT_NE(parent->GetFirstElement(), nullptr);
 
   parent->AddAttribute("test", "string", "foo", false, "foo description");
   ASSERT_EQ(parent->GetAttributeCount(), 1UL);
 
-  std::string stringval = parent->ToString("myprefix");
-  ASSERT_EQ(stringval, "myprefix< test='foo'>\nmyprefix  </>\nmyprefix</>\n");
+  // attribute won't print unless it has been set
+  EXPECT_FALSE(parent->GetAttributeSet("test"));
+  EXPECT_EQ(parent->ToString("<!-- prefix -->"),
+    "<!-- prefix --><parent>\n"
+    "<!-- prefix -->  <child/>\n"
+    "<!-- prefix --></parent>\n");
+
+  sdf::ParamPtr test = parent->GetAttribute("test");
+  ASSERT_NE(nullptr, test);
+  EXPECT_FALSE(test->GetSet());
+  EXPECT_TRUE(test->SetFromString("foo"));
+  EXPECT_TRUE(test->GetSet());
+  EXPECT_TRUE(parent->GetAttributeSet("test"));
+
+  EXPECT_EQ(parent->ToString("<!-- prefix -->"),
+    "<!-- prefix --><parent test='foo'>\n"
+    "<!-- prefix -->  <child/>\n"
+    "<!-- prefix --></parent>\n");
 }
 
 /////////////////////////////////////////////////
@@ -316,8 +335,18 @@ TEST(Element, ToStringValue)
 
   parent->AddValue("string", "val", false, "val description");
 
-  std::string stringval = parent->ToString("myprefix");
-  ASSERT_EQ(stringval, "myprefix< test='foo'>val</>\n");
+  EXPECT_FALSE(parent->GetAttributeSet("test"));
+  EXPECT_EQ(parent->ToString("myprefix"),
+            "myprefix<>val</>\n");
+
+  sdf::ParamPtr test = parent->GetAttribute("test");
+  ASSERT_NE(nullptr, test);
+  EXPECT_FALSE(test->GetSet());
+  EXPECT_TRUE(test->SetFromString("foo"));
+  EXPECT_TRUE(test->GetSet());
+  EXPECT_TRUE(parent->GetAttributeSet("test"));
+  EXPECT_EQ(parent->ToString("myprefix"),
+            "myprefix< test='foo'>val</>\n");
 }
 
 /////////////////////////////////////////////////