+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### SDFormat 8.X.X (201X-XX-XX)
 
+1. Element: don't print unset attributes.
+    * [Pull request 571](https://bitbucket.org/osrf/sdformat/pull-requests/571)
+
 1. Set camera intrinsics axis skew (s) default value to 0
     * [Pull request 504](https://bitbucket.org/osrf/sdformat/pull-requests/504)